junit tests

+++ item/item-api-domain/src/main/java/com/nextuple/item/domain/events/ItemMasterEvent.java
@@ -37,7 +37,7 @@ public class ItemMasterEvent implements Serializable {
   private String volumeUom;
   private Double weight;
   private String weightUom;
-  private Long processingTime;
+  private Double processingTime;
   private String cost;
   private Boolean isHazmat;
   private String shortDescription;
